requirement.txt
absl-py==1.1.0
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
astroid==2.11.7
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click
crypto==1.4.1
cssselect==1.1.0
cssutils
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.13
dill==0.3.5.1
docker-pycreds==0.4.0
entrypoints==0.4
fastjsonschema==2.16.1
flake8
flask
fonttools==4.33.3
future==0.18.2
gevent==21.12.0
gitdb==4.0.9
gitpython==3.1.27
google-auth-oauthlib==0.4.6
google-auth==2.9.0
greenlet==1.1.2
grpcio==1.47.0
idna==3.3
imageio==2.19.3
imgaug==0.4.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython-genutils==0.2.0
ipython==7.34.0
ipywidgets==7.7.1
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
jinja2==3.1.2
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter==1.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
lazy-object-proxy==1.7.1
lxml==4.9.1
markdown
markupsafe==2.1.1
matplotlib-inline==0.1.3
matplotlib==3.5.2
mccabe==0.7.0
mistune==0.8.4
naked==0.1.31
nbclient==0.6.6
nbconvert==6.5.1
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.12
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.1.1
pip==21.2.2
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
premailer==3.10.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.9.0
pycparser==2.21
pycrypto==2.6.1
pyflakes==2.5.0
pygments==2.12.0
pylatex==1.4.1
pylint==2.14.5
pynvml==11.4.1
pyparsing==3.0.9
pyqt5-qt5==5.15.2
pyqt5-sip==12.11.0
pyqt5==5.15.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pywavelets==1.3.0
pyyaml==6.0
pyzmq==23.2.0
qtconsole==5.3.1
qtpy==2.2.0
redis==4.3.4
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.8
scikit-image==0.19.3
scipy==1.7.3
screen==1.0.1
seaborn==0.11.2
send2trash==1.8.0
sentry-sdk==1.6.0
setproctitle==1.2.3
setuptools==59.5.0
shapely==1.8.2
shellescape==3.8.1
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.9.1
terminado==0.15.0
terminaltables==3.1.10
thop==0.1.0.post2207010342
tifffile==2021.11.2
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
torch==1.10.1
torchsummary==1.5.1
torchvision==0.11.2
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typed-ast==1.5.4
typing-extensions==4.2.0
urllib3==1.26.9
wandb==0.12.21
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.1.2
wheel==0.37.1
widgetsnbextension==3.6.1
wrapt==1.14.1
yagmail==0.15.280
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0